Address issue of some zipped directories not having '.zip' extension
When storing directories in FIRE they are zipped and have to be
accessed as a zip file. However, as of the time of this commit the
biostudies API does not append some directory names and corresponding
uri with .zip. Additionally, the API does not have a flag to allow
identification of directories.

This commit addresses these issues by checking if fileref uris end with
a known extension. If not a head request to the uri with '.zip' is made
and the uri and name are appended with '.zip' if the response was 200.