Disable exit on error for rule: get_image_reps_to_convert
For some reason (probably the pipes) one of the bash lines in the
get_image_reps_to_convert rule causes snakemake to exit as though there
was a bash error - even though no error occurred. This was observed in
about 10 out of 52 studies processed in jan-mar 2023 workflow e.g.
S-BIAD686. Disabling exit on error seems to fix this.