Fix issues from PR #32 review. Mainly - use 'type' from biostudies api
After review of PR #32, the consensus was to use the 'type' attribute
of a file reference returned by the biostudies api when ingesting from
biostudies - instead of attempting to classify file references as zipped
directories/zarrs/archives. This commit addresses the necessary changes
in the ingest_from_biostudies.py script and downstream files.