Fix bug causing some images to be missed when converting to zarr
The regex used to select filenames in the 'aggregate_converted_im_reps'
function was causing some files to be skipped and this propagated to
the 'convert_to_zarr_and_upload' stage. This bug was fixed by using
a python lambda expression on the filenames to do the selection
instead of a regex.